MICNO-730: rename manager class & move rollback code to domain module
