+++ job_executor/manager/__init__.py
@@ -1,9 +1,10 @@
+from job_executor.adapter import job_service
 from job_executor.model.worker import Worker
 
 from typing import List
 
 
-class ManagerState:
+class Manager:
     def __init__(self, max_workers, max_bytes_all_workers):
         """
         :param default_max_workers: The maximum number of workers
@@ -61,3 +62,23 @@ class ManagerState:
         self.workers = [
             worker for worker in self.workers if worker.job_id != job_id
         ]
+
+    def clean_up_after_dead_workers(self) -> None:
+        dead_workers = self.dead_workers
+        if len(dead_workers) > 0:
+            in_progress_jobs = job_service.get_jobs(ignore_completed=True)
+            for dead_worker in dead_workers:
+                job = next(
+                    (
+                        job
+                        for job in in_progress_jobs
+                        if dead_worker.job_id == job.job_id
+                    ),
+                    None,  # not found in in_progress => completed or failed
+                )
+                if job and job.status not in ["queued", "built"]:
+                    logger.warning(
+                        f"Worker died and did not finish job {job.job_id}"
+                    )
+                    fix_interrupted_job(job)
+                self.unregister_job(dead_worker.job_id)
